refactored entire front-end code and improved backend

+++ API/Entities/Order.cs
@@ -9,7 +9,7 @@ namespace API.Entities
     {
     }
 
-    public Order(List<Product> orderItems ,string email, string address, string city,  string country, string zipcode, string firstName, string lastName, string telephone, string orderNotes, decimal total)
+    public Order(List<Product> orderItems ,string email, string address, string city,  string country, string zipcode, string firstName, string lastName, string telephone, string orderNotes, int total)
     {
       Email = email;
       Address = address;
@@ -38,7 +38,7 @@ namespace API.Entities
     public string OrderNotes { get; set; }
     public DateTimeOffset OrderDate {get; set;} = DateTimeOffset.Now;
     public List<Product> OrderItems {get; set;}
-    public decimal Total { get; set; }
+    public int Total { get; set; }
     // public string PaymentIntentId { get; set; }
